mysite/requirements.txt
asgiref==3.2.10
Django==3.1
pytz==2020.1
sqlparse==0.3.1

djangorestframework==3.11.0
djoser==2.0.3
django-cors-headers==3.1.1

